Refactor audio file helpers into utils
