+++ src/audio/__init__.py
@@ -1,10 +1,10 @@
-from .noise_file import (
+from .utils.noise_file import (
     NoiseParams,
     save_noise_params,
     load_noise_params,
     NOISE_FILE_EXTENSION,
 )
-from .voice_file import (
+from .utils.voice_file import (
     VoicePreset,
     save_voice_preset,
     load_voice_preset,
